{"schema":"decodex/commit/1","summary":"Split Decodex app account store modules","authority":"manual"}
